requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0